Replace iterator functions
