+++ lib/ffi/Cargo.toml
@@ -9,6 +9,7 @@ authors = ["Yota Toyama <raviqqe@gmail.com>"]
 repository = "https://github.com/pen-lang/pen"
 
 [dependencies]
+async-stream = "0.3.3"
 futures = "0.3"
 pen-ffi-macro = { version = "0.1", path = "../ffi-macro" }
