+++ packages/ffi/list.pen
@@ -1,13 +1,13 @@
 type firstRest {
   ok boolean
-  first any
+  first \() any
   rest [any]
 }
 
 foreign _pen_ffi_list_first_rest = \(xs [any]) firstRest {
   if [x, ...xs] = xs {
-    firstRest{ok: true, first: x(), rest: xs}
+    firstRest{ok: true, first: x, rest: xs}
   } else {
-    firstRest{ok: false, first: none, rest: [any]}
+    firstRest{ok: false, first: \() any { none }, rest: [any]}
   }
 }
